1.更改了购物车所有接口的入参校验 2.修改了serviceImpl中的商品信息校验

+++ src/main/java/so/sao/shop/supplier/web/app/AppCartController.java
@@ -10,12 +10,12 @@ import so.sao.shop.supplier.config.Constant;
 import so.sao.shop.supplier.pojo.Result;
 import so.sao.shop.supplier.pojo.input.AppCartItemBatchInput;
 import so.sao.shop.supplier.pojo.input.AppCartItemInput;
+import so.sao.shop.supplier.pojo.input.AppCartItemSaveInput;
 import so.sao.shop.supplier.pojo.output.AppCartItemOut;
 import so.sao.shop.supplier.service.app.AppCartService;
+import so.sao.shop.supplier.util.Ognl;
 
 import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
 import java.util.List;
 import java.util.Map;
 
@@ -42,8 +42,13 @@ public class AppCartController {
      */
     @ApiOperation(value = "根据购物车ID删除购物车记录", notes = "根据购物车ID删除购物车记录")
     @GetMapping(value = "/cartitem/delete")
-    public Result deleteCartItem(@NotNull(message = "购物车记录ID不能为空") Long cartitemId,
-                                 @NotNull(message = "用户ID不能为空") Long userId) throws Exception {
+    public Result deleteCartItem(Long cartitemId, Long userId) throws Exception {
+        if (Ognl.isNull(cartitemId)) {
+            return Result.fail("购物车记录ID不能为空");
+        }
+        if (Ognl.isNull(userId)) {
+            return Result.fail("用户ID不能为空");
+        }
         // 删除购物车记录
         Boolean flag = cartService.deleteCartItemById(cartitemId, userId);
         // 返回结果
@@ -98,19 +103,17 @@ public class AppCartController {
     /**
      * 向购物车添加购物记录
      *
-     * @param number
-     * @param userId
+     * @param input
      * @return
      * @throws Exception
      */
     @ApiOperation(value = "向购物车添加购物记录", notes = "向购物车添加购物记录")
     @PostMapping(value = "/cartitem/save")
-    public Result createCartItems(@NotNull(message = "商品ID不能为空") Long commodityId,
-                                  @NotNull(message = "购物车添加商品数量不能为空")
-                                  @Pattern(regexp = "^[1-9][0-9]*$", message = "购物车添加商品数量有误") Integer number,
-                                  @NotNull(message = "用户ID不能为空") Long userId) throws Exception {
+    public Result createCartItems(@RequestBody @Valid AppCartItemSaveInput input) throws Exception {
+        // 转化为Integer
+        Integer newNumber = Integer.valueOf(input.getNumber());
         // 插入数据
-        Map<String, Object> map = cartService.saveCartItem(commodityId, number, userId);
+        Map<String, Object> map = cartService.saveCartItem(input.getCommodityId(), newNumber, input.getUserId());
         // 获取提示信息
         String msg = (String) map.get("msg");
         // 获取信息码
@@ -131,7 +134,10 @@ public class AppCartController {
      */
     @ApiOperation(value = "根据用户id获取用户购物车信息", notes = "根据用户id获取用户购物车信息【负责人：王翼云】")
     @GetMapping(value = "/cartitem/select")
-    public Result getCartItemsByUser(@NotNull(message = "用户ID不能为空") Long userId) throws Exception {
+    public Result getCartItemsByUser(Long userId) throws Exception {
+        if (Ognl.isNull(userId)) {
+            return Result.fail("用户ID不能为空");
+        }
         // 查询数据
         Map<String, Object> map = cartService.findCartItemsByUserId(userId);
         // 获取信息码
